Fix: merge webpack config from user (#136)
* fix: merge webpack config from user


* chore: add function descripttion


* chore: remove deep merge


* fix: remove deep merge export


* fix: correct order of spread operators


* chore: spelling correction


* chore: add changeset


* fix: correct order of spread operators


* chore: add reminder comment


---------
