+++ .changeset/odd-fans-hope.md
@@ -1,0 +1,5 @@
+---
+"@snapwp/next": patch
+---
+
+fix: allow users to extend webpack config.
